// src/grid/types.ts
export type ViewportType = 'xs' | 'sm' | 'md' | 'lg';

export type RowKeyType =
  | 'start'
  | 'center'
  | 'end'
  | 'top'
  | 'middle'
  | 'bottom'
  | 'around'
  | 'between';

export type ColKeyType =
  | 'xs'
  | 'sm'
  | 'md'
  | 'lg'
  | 'xl'
  | 'xsOffset'
  | 'smOffset'
  | 'mdOffset'
  | 'lgOffset'
  | 'xlOffset';